+++ runtime/io.js
@@ -277,9 +277,10 @@ function caml_ml_input_block (chanid, ba, i, l) {
 //Provides: caml_input_value
 //Requires: caml_marshal_data_size, caml_input_value_from_bytes, caml_create_bytes, caml_ml_channels, caml_bytes_of_array
 //Requires: caml_refill, caml_failwith, caml_raise_end_of_file
+//Requires: caml_marshal_header_size
 function caml_input_value (chanid) {
   var chan = caml_ml_channels[chanid];
-  var header = new Uint8Array(20);
+  var header = new Uint8Array(caml_marshal_header_size);
   function block(buffer, offset, n) {
     var r = 0;
     while(r < n){
@@ -296,15 +297,15 @@ function caml_input_value (chanid) {
     }
     return r;
   }
-  var r = block(header, 0, 20);
+  var r = block(header, 0, caml_marshal_header_size);
   if(r == 0)
     caml_raise_end_of_file();
-  else if (r < 20)
+  else if (r < caml_marshal_header_size)
     caml_failwith("input_value: truncated object");
   var len = caml_marshal_data_size (caml_bytes_of_array(header), 0);
-  var buf = new Uint8Array(len + 20);
+  var buf = new Uint8Array(len + caml_marshal_header_size);
   buf.set(header,0);
-  var r = block(buf, 20, len)
+  var r = block(buf, caml_marshal_header_size, len)
   if(r < len)
     caml_failwith("input_value: truncated object " + r + "  " + len);
   var offset = [0];